move BaseEndpoints to common.api package
